Skip tests with StopIteration under python 3.7+
PEP 479 treats the 'raise StopIteration' in a generator function as an
error, and translates that StopIteration to RuntimeError. (More
precisely speaking, the exception may be raised explicitly or
implicitly by a nested call. It becomes a RuntimeError if this
exception is allowed to terminate the function. If it is caught before
that, that's OK.)

At the same time, python2 does not allow *anything* after 'return' in
a generator function, raising a SyntaxError while parsing the code.

Thus, even though it's simple to write code which runs under
python <= 3.6 (just 'raise StopIteration(args)'), and it's simple to
write equivalent code that runs under python >= 3.3 (just 'return args'),
I don't see a way to write code that would pass on both.

Let's just skip the test under python3.7+, and when support for
python2.7 is dropped, we can use 'return …', and remove the skip mark.